% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pd.R
\name{violin_immune}
\alias{violin_immune}
\title{Title}
\usage{
violin_immune(
  rt,
  group_list,
  Control = "CT",
  Treat = "PC",
  f = "./Figure5/vioplot.pdf"
)
}
\arguments{
\item{f}{}
}
\description{
Title
}
